Fix: 로그인 및 회원가입 수정

+++ src/main/java/com/ssafy/enjoytrip/user/service/UserServiceImpl.java
@@ -11,6 +11,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import java.util.HashMap;
+import java.util.Map;
 
 @Slf4j
 @Service
@@ -47,7 +49,6 @@ public class UserServiceImpl implements UserService {
         checkDupEmail(user.getEmail());
         checkDupNickname(user.getNickname());
         String encodedPassword = passwordEncoder.encode(user.getPassword());
-        System.out.println(encodedPassword);
         user.setPassword(encodedPassword);
         userDao.insert(user);
     }
@@ -61,7 +62,24 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     @Override
-    public int dropOutById(int id) {
+    public int modifyPwd(int id, String curpwd, String newpwd) {
+        User user = userDao.findById(id);
+        if (!passwordCheck(curpwd, user.getPassword())) {
+            throw new InvalidPasswordException("로그인 실패");
+        }
+        String encodedPassword = passwordEncoder.encode(newpwd);
+        User newUserPwd = new User(encodedPassword);
+        newUserPwd.setId(id);
+        return userDao.updatePwd(newUserPwd);
+    }
+
+    @Transactional
+    @Override
+    public int dropOutById(int id, String pwd) {
+        User user = userDao.findById(id);
+        if (!passwordCheck(pwd, user.getPassword())) {
+            throw new InvalidPasswordException("비밀번호 불일치");
+        }
         return userDao.deleteById(id);
     }
 
@@ -87,5 +105,25 @@ public class UserServiceImpl implements UserService {
         return userDao.findById(id);
     }
 
+    @Override
+    public void saveRefreshToken(int userid, String refreshToken) throws Exception {
+        Map<String, Object> map = new HashMap<>();
+        map.put("id", userid);
+        map.put("token", refreshToken);
+        userDao.saveRefreshToken(map);
+    }
+
+    @Override
+    public Object getRefreshToken(int userid) throws Exception {
+        return userDao.getRefreshToken(userid);
+    }
+
+    @Override
+    public void deleteRefreshToken(int userid) throws Exception {
+        Map<String, Object> map = new HashMap<>();
+        map.put("userid", userid);
+        map.put("token", null);
+        userDao.deleteRefreshToken(map);
+    }
 
 }
